feat(create): add Showcase preview view (landing cards grid)
Adds a "Showcase" option to the /create preview selector that renders the
landing page's masonry of real-world cards inside the preview iframe, themed
live by the user's design system — so the whole system can be judged at a
glance on realistic UI rather than one component at a time.

- Extract the shared card masonry into `CardsGrid`
  (components/marketing/showcase/cards-grid.tsx); the landing `cards.tsx` and
  the new preview view now render the exact same grid. Column count and width
  vary by context via a `className` merged over the canonical defaults.
- Add the `showcase` preview group-example so the iframe resolves
  `preview=showcase` through the existing GroupExamplesIndex path — no route
  changes needed.
- Surface "Showcase" at the top of the preview Select in the customizer panel.
- Regenerate examples.tsx (+showcase) and the v0 showcase-bundle (cards.tsx
  import closure now includes cards-grid.tsx).
